Fix symlink problems.
When a symlink of app.sh is moved to another directory, the symlink won't work.
Because app.sh finds its placed directory with `dirname $0`, symlinked file couldn't find app.sh's directory.
I modified `dirname $0` to `dirname $(realpath $(which $0))`. This works better.
However if a content of $0 is alias, this won't works.